+++ offchain/advance-runner/src/config.rs
@@ -14,8 +14,8 @@ use clap::Parser;
 use snafu::{ResultExt, Snafu};
 use std::time::Duration;
 
-use crate::server_manager::config::ServerManagerCLIConfig;
-pub use crate::server_manager::config::ServerManagerConfig;
+use crate::server_manager::ServerManagerCLIConfig;
+pub use crate::server_manager::ServerManagerConfig;
 pub use crate::snapshot::config::{FSManagerConfig, SnapshotConfig};
 use crate::snapshot::config::{SnapshotCLIConfig, SnapshotConfigError};
 pub use http_health_check::HealthCheckConfig;