+++ offchain/advance-runner/src/server_manager/claim.rs
@@ -14,33 +14,41 @@ use rollups_events::{Hash, HASH_SIZE};
 use sha3::{Digest, Keccak256};
 
 pub fn compute_claim_hash(
-    machine_state_hash: &[u8],
-    vouchers_metadata_hash: &[u8],
-    notices_metadata_hash: &[u8],
+    machine_state_hash: &Hash,
+    vouchers_metadata_hash: &Hash,
+    notices_metadata_hash: &Hash,
 ) -> Hash {
     let mut hasher = Keccak256::new();
-    hasher.update(machine_state_hash);
-    hasher.update(vouchers_metadata_hash);
-    hasher.update(notices_metadata_hash);
+    hasher.update(machine_state_hash.inner());
+    hasher.update(vouchers_metadata_hash.inner());
+    hasher.update(notices_metadata_hash.inner());
     let data: [u8; HASH_SIZE] = hasher.finalize().into();
     Hash::new(data)
 }
 
 #[cfg(test)]
 mod tests {
-    use super::compute_claim_hash;
+    use super::{compute_claim_hash, Hash};
 
     #[test_log::test]
     fn test_claim_hash() {
-        let hash = hex::decode(
-            "ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff",
-        )
-        .unwrap();
+        let hash = Hash::new(
+            hex::decode(
+                "ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff",
+            )
+            .unwrap()
+            .try_into()
+            .unwrap()
+        );
         let claim = compute_claim_hash(&hash, &hash, &hash);
-        let expected = hex::decode(
-            "8590bbc3ea43e28e8624fb1a2d59aaca701a5517e08511c4a14d9037de6f6086",
-        )
-        .unwrap();
-        assert_eq!(&expected, &claim.inner());
+        let expected = Hash::new(
+            hex::decode(
+                "8590bbc3ea43e28e8624fb1a2d59aaca701a5517e08511c4a14d9037de6f6086",
+            )
+            .unwrap()
+            .try_into()
+            .unwrap()
+        );
+        assert_eq!(expected, claim);
     }
 }