+++ offchain/test-fixtures/src/machine_snapshots.rs
@@ -19,7 +19,8 @@ use crate::docker_cli;
 
 const TAG: &str = "cartesi/test-machine-snapshot";
 const DOCKERFILE: &str = "../test-fixtures/docker/machine_snapshot.Dockerfile";
-const CONTAINER_SNAPSHOT_DIR: &str = "/opt/cartesi/share/dapp-bin/0";
+const CONTAINER_SNAPSHOT_DIR: &str = "/opt/cartesi/share/dapp-bin";
+const SNAPSHOT_NAME: &str = "0_0";
 
 pub struct MachineSnapshotsFixture {
     dir: TempDir,
@@ -33,11 +34,15 @@ impl MachineSnapshotsFixture {
         let dir = tempfile::tempdir().expect("failed to create temp dir");
         docker_cli::build(DOCKERFILE, TAG, &[]);
         let id = docker_cli::create(TAG);
-        let from = format!("{}:{}", id, CONTAINER_SNAPSHOT_DIR);
-        docker_cli::cp(&from, dir.path().to_str().unwrap());
+        let from_container = format!("{}:{}", id, CONTAINER_SNAPSHOT_DIR);
+        let to_host = dir.path().join(SNAPSHOT_NAME);
+        docker_cli::cp(&from_container, to_host.to_str().unwrap());
         docker_cli::rm(&id);
-        unix::fs::symlink(dir.path().join("0"), dir.path().join("latest"))
-            .expect("failed to create latest link");
+        unix::fs::symlink(
+            dir.path().join(SNAPSHOT_NAME),
+            dir.path().join("latest"),
+        )
+        .expect("failed to create latest link");
         Self { dir }
     }
 
@@ -48,9 +53,19 @@ impl MachineSnapshotsFixture {
 
     /// Check whether the given snapshot is the latest
     #[tracing::instrument(level = "trace", skip_all)]
-    pub fn assert_latest_snapshot(&self, epoch_index: u64) {
-        tracing::trace!(epoch_index, "checking the latest snapshot");
-        let snapshot = self.path().join(epoch_index.to_string());
+    pub fn assert_latest_snapshot(
+        &self,
+        epoch_index: u64,
+        processed_input_count: u64,
+    ) {
+        tracing::trace!(
+            epoch_index,
+            processed_input_count,
+            "checking the latest snapshot"
+        );
+        let snapshot_name =
+            format!("{}_{}", epoch_index, processed_input_count);
+        let snapshot = self.path().join(snapshot_name);
         assert!(snapshot.is_dir(), "snapshot not found");
         let latest = self.path().join("latest");
         assert!(latest.is_symlink(), "latest link not found");